==> FirstTrifo/5_FirstTrifo_Iv2_Calc5Graphs.Rmd <==
---
title: "First Trifoliate appearance I2 graphs with Mean Air Temp"
author: "CT"
date: "17.08.2017"
output: 
  word_document: default
  pdf_document: default
  html_document: default
---

This script does this:
-get dataframe produced in Calc4_2(df_MeanTempFirstTrifo.txt) which contains all elements to produce graphs! 
-graph y=Days to First trifoliate ; x= mean air Temp and soil Temp
-graph y= 1/days to First trifoliates ; x= mean air Temp and soil Temp 
-produces the linear equations leaf appearance  as a function of temperature 



```{r loadLibraries, warning=FALSE}
library(tidyr)
library(dplyr)
library(ggplot2)
library(agricolae)
library(knitr)
library(lubridate)
library(mosaic)
library(lattice)
library (gridExtra)
```

Read file and correct formats 
```{r loadFile}
# Get dataframe from previous script (written table )
setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\FirstTrifo")
getwd()

#create file
#df_G <- read.table("df_DASFirstTrifo_MeanAirT.txt",header=TRUE)
df_G <- read.table("df_MeanTempFirstTrifo.txt",header=TRUE)

"df_MeanTempFirstTrifo.txt"


```

```{r}
df_G$DAS1T <- as.numeric(df_G$DAS1T)
df_G$SowingDate <- ymd(df_G$SowingDate )
df_G$Cultivar <- as.factor(df_G$Cultivar)
df_G$Plot <- as.factor(df_G$Plot)

#check data 
summary(df_G)



# this is to "clean previous dfs : df_emerg4 <- NULL
```


Bring data Rate from emergence (not from sowing !!)

```{r}

df_trifEmerg <- read.table("RateFirstTrifoFromEmer.txt",header=TRUE)



df_trifEmerg<-df_trifEmerg%>%
  dplyr::select (Plot,Rate,DASfromEmerg )%>%
  mutate(RateFromEmerg=Rate)%>%
  dplyr::select(-Rate)
  
  str(df_trifEmerg)
  
  ##then merge dataframes
  
 Merg<- merge(df_G,df_trifEmerg, by="Plot")
  
 ##exclude data outliers 
 
  Merg<-Merg%>%
    subset(RateFromEmerg>0)

```



```{r}
df_mean<-Merg%>%
  group_by(Cultivar,SowTreat)%>%
   summarise_all(funs(mean))



df_meanSow <-Merg %>%
  group_by(SowTreat)%>%
   summarise_all(funs(mean))

```



Graph DAS1T (days to first trifo )  and rate (first trifo) against meanSoil Temp at trifoliate appearance  

```{r, fig.height=5, fig.width=10, echo = FALSE}

str(df_mean)

fontSize <- 12.0


  g1 <- df_mean %>%
  #ggplot(aes(x=  meanTempSoil_1T, y=DAS1T, shape=SowTreat))  +
    ggplot(aes(x=  meanTempSoil_1T, y=DASfromEmerg, shape=SowTreat))  +
  #geom_point(alpha=0.8) +
  geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  #geom_smooth()+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Soil~temperature~~"("^"o"*"C)")) +  
  ylab("Days to first trifoliate  (days)") +
    facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,40))
  
   g2 <- df_mean %>%
  #ggplot(aes(x=  meanTempSoil_1T, y=DAS1T, shape=SowTreat))  +
    ggplot(aes(x=  meanTempSoil_1T, y=RateFromEmerg, shape=SowTreat))  +
  #geom_point(alpha=0.8) +
  geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  #geom_smooth()+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Soil~temperature~~"("^"o"*"C)")) +  
  #ylab("Days to first trifoliate  (days)") +
     ylab(" First Trifoliate  Rate (1/ Days) ") +
    facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,0.2))

#
  
  #grid.arrange(g1, g2)
# 

 g <-grid.arrange(g1, g2)
  
  #ggsave(file="plot_DAS_Rate_FirstTrifoIversenSoilTemp.png", width=12, height=6, dpi=400,g)
 ggsave(file="plot_DAS_Rate_FirstTrifoIvEmerg_SoilTemp.tiff", width=12, height=6, dpi=400,g)

```



then do the graph with air temperature 


```{r,fig.height=5, fig.width=10, echo = FALSE}

 g3 <- df_mean %>%
  #ggplot(aes(x=  meanTempAir_1T, y=DAS1T, shape=SowTreat))  +
  ggplot(aes(x=  meanTempAir_1T, y=DASfromEmerg, shape=SowTreat))  +
  #geom_point(alpha=0.8) +
  geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  #geom_smooth()+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Air~temperature~~"("^"o"*"C)")) +  
  ylab("Days to first trifoliate  (days)") +
    facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,40))

# 1/DAS1T
  g4 <- df_mean %>%
  ggplot(aes(x= meanTempAir_1T, y=RateFromEmerg, shape=SowTreat))  +
    
  #geom_point(alpha=0.8) +
  geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  #geom_smooth(colour="red")+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Air~temperature~~"("^"o"*"C)")) + 
  ylab(" First Trifoliate  Rate (1/ Days) ") +
  facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
 scale_x_continuous(limits = c(0,25))
 scale_y_continuous(limits = c(0,0.5))
  
   g <-grid.arrange(g3, g4)
  
  #ggsave(file="plot_DAS_Rate_FirstTrifoIvAirTemp.png", width=12, height=6, dpi=400,g)




```



Then get the equations from the whole temperature range  in all sowing dates  



use the means for each sowtreat 



calculate mean values for each cultivar and sow Treat

```{r}
df_mean<-df_G%>%
  group_by(Cultivar,SowTreat)%>%
   summarise_all(funs(mean,sd))
```

```{r, fig.height=5, fig.width=12, echo = FALSE}
#DAS50E
#
  g3 <- df_mean %>%
  #filter(Cultivar=="Monti"| Cultivar=="Antas")%>%
 # filter (SowTreat== "S3"| SowTreat== "S4"|SowTreat== "S5"|SowTreat== "S6")%>%
    #group_by(Cultivar,SowTreat)%>%
    #summarise_all(funs(mean))
  #ggplot(aes(x=MeanAirTempTrifo_mean, y=DAS1T_mean, colour=SowTreat))  +
  ggplot(aes(x=MeanAirTempTrifo_mean, y=DAS1T_mean)) +
    geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  
  #geom_errorbar(aes(ymin=DAS5T_mean-DAS5T_sd/2,
  #ymax=DAS5T_mean+DAS5T_sd/2),width=3,colour="grey")+
  #geom_errorbar(aes(ymin=2-DAS5T_sd/2,
  #ymax=2+DAS5T_sd/2),width=2,colour="grey")+
 # geom_point(alpha=0.8) +
  #geom_smooth()+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Air~temperaturee~~"("^"o"*"C)")) +  
  ylab("Days to First trifoliate (days)") +
    facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
  scale_x_continuous(limits = c(0,20))+
  scale_y_continuous(limits = c(0,90))

# 1/DAS50E
  g4 <- df_mean %>%
   #filter(Cultivar=="Monti"| Cultivar=="Antas")%>%
  #filter (SowTreat== "S3"| SowTreat== "S4"|SowTreat== "S5"|SowTreat== "S6")%>%
  #ggplot(aes(x=MeanAirTempTrifo_mean, y=Rate_mean, colour=SowTreat))  +
    ggplot(aes(x=MeanAirTempTrifo_mean, y=Rate_mean)) +
    geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  #geom_point(alpha=0.8) +
  #geom_smooth(colour="red")+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Air~temperature~~"("^"o"*"C)")) + 
  ylab(" Rate (1/ Days) to First trifoliate") +
  facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
 scale_x_continuous(limits = c(0,20))+
  scale_y_continuous(limits = c(0,0.08))
  grid.arrange(g3, g4)
  
  
  
```


Plot days to first trifoliate agains temperature 

```{r}

require(ggplot2)
require(plyr)
require(dplyr)

lm_eqn = function(df){
  m = lm(y ~ x, df);
  eq <- substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2, 
                   list(a = format(coef(m)[1], digits = 2), 
                        b = format(coef(m)[2], digits = 2), 
                        r2 = format(summary(m)$r.squared, digits = 3)))
  as.character(as.expression(eq));                 
}


df <- df_G %>%
  dplyr::select(SowTreat,Cultivar, Plot, MeanAirTempTrifo, DAS1T) %>%
  group_by(SowTreat,Cultivar) %>%
  summarise_all(funs(mean)) %>%
  mutate(y=DAS1T,x=MeanAirTempTrifo) %>%
  dplyr::select(-MeanAirTempTrifo, -DAS1T)

eq <- ddply(df,.(Cultivar),lm_eqn)

yx<-quantile(df$y,0.75)
xn<- quantile(df$x,0.25)

#get statistical summary here 
summary(lm(y ~ x, df))

p <- ggplot(data = df, aes(x = x, y = y)) +
  geom_smooth(method = "lm", se=FALSE, color="black", formula = y ~ x) +
  geom_point()+
  theme_bw()+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,60))

p1 <- p + geom_text(data=eq,aes(x = xn, y = yx,label=V1), parse = TRUE, inherit.aes=FALSE) + 
  facet_grid(Cultivar~.,scales="free")
  
p1

```



Plot Rate of emergence against temperature 


```{r, fig.height=7,fig.width=10}
require(ggplot2)
require(plyr)
require(dplyr)

lm_eqn = function(df){
  m = lm(y ~ x, df);
  eq <- substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2, 
                   list(a = format(coef(m)[1], digits = 2), 
                        b = format(coef(m)[2], digits = 2), 
                        r2 = format(summary(m)$r.squared, digits = 3)))
  as.character(as.expression(eq));                 
}


df <- df_G %>%
  dplyr::select(SowTreat,Cultivar, Plot, MeanAirTempTrifo, Rate) %>%
  group_by(SowTreat,Cultivar) %>%
  summarise_all(funs(mean)) %>%
  mutate(y=Rate,x=MeanAirTempTrifo) %>%
  dplyr::select(-MeanAirTempTrifo, -Rate)

eq <- ddply(df,.(Cultivar),lm_eqn)

yx<-quantile(df$y,0.75)
xn<- quantile(df$x,0.25)

#get statistical summary here 
summary(lm(y ~ x, df))

p <- ggplot(data = df, aes(x = x, y = y)) +
  geom_smooth(method = "lm", se=FALSE, color="grey",linetype=2, formula = y ~ x) +
  geom_point()+
  theme_bw()+
  scale_x_continuous(limits = c(0,18))+
  scale_y_continuous(limits = c(0,0.08))

p1 <- p + geom_text(data=eq,aes(x = xn, y = yx,label=V1), parse = TRUE, inherit.aes=FALSE) + 
  facet_grid(.~Cultivar,scales="free",ncol(3))+
  ylab("Rate")
  
p1
```

Do it individually for each cultivar
```{r}

require(ggplot2)
require(plyr)
require(dplyr)

lm_eqn = function(df){
  m = lm(y ~ x, df);
  eq <- substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2, 
                   list(a = format(coef(m)[1], digits = 2), 
                        b = format(coef(m)[2], digits = 2), 
                        r2 = format(summary(m)$r.squared, digits = 3)))
  as.character(as.expression(eq));                 
}


df <- df_G %>%
  dplyr::select(SowTreat,Cultivar, Plot, MeanAirTempTrifo, Rate) %>%
  group_by(SowTreat,Cultivar) %>%
  filter(Cultivar=='Denmark')%>%
  summarise_each(funs(mean)) %>%
  mutate(y=Rate,x=MeanAirTempTrifo) %>%
  dplyr::select(-MeanAirTempTrifo, -Rate)

eq <- ddply(df,.(Cultivar),lm_eqn)

yx<-quantile(df$y,0.75)
xn<- quantile(df$x,0.25)

#get statistical summary here 
summary(lm(y ~ x, df))

p <- ggplot(data = df, aes(x = x, y = y)) +
  geom_smooth(method = "lm", se=FALSE, color="black", formula = y ~ x) +
  geom_point()+
  theme_bw()+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,0.045))

p1 <- p + geom_text(data=eq,aes(x = xn, y = yx,label=V1), parse = TRUE, inherit.aes=FALSE) + 
  facet_grid(Cultivar~.,scales="free")
  
p1

```

  
  
  get the rate means for each sowing date 
  
```{r}

df_meanSowT <-df_mean%>%
  group_by(SowTreat)%>%
   summarise_all(funs(mean,sd))

  


```
