==> FirstTrifo/6_First_TrifoIv2_Calc6_ThermalTime.Rmd <==
---
title: "First Trifoliates Iversen "
author: "CT"
date: "17.08.2017"
output: html_document
---

Purpose:

This script does this:

- get the dataframe with sowing dates and date to 50% first trifo   
- get the dataframe with soil temperature (tmax and t min) 
- merge the two dataframes to calculate the Thermal time ( using the 8 X 3 formula )to  first trifo appearance for each specific cultivars using Tb=0
  - calculates the DAS and TTFirstTrifo from emergence 
 
```{r loadLibraries}
library(tidyr)
library (dplyr)
library (ggplot2)
library(agricolae)
library (lubridate)
library(mosaic)
library(lattice)
```

Get data from Frist Trifo folder  


```{r}

setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\FirstTrifo")
getwd()
#create file

df_trif <- read.table("RateFirstTrifo.txt",header=TRUE)
#adjust formats 
df_trif$SowingDate <- ymd(df_trif$SowingDate )
df_trif$Block <- as.factor(df_trif$Block)
df_trif$Plot <- as.factor(df_trif$Plot)
df_trif$DAS1T <- as.numeric(df_trif$DAS1T)

```

Create df_MD to workout dates to % trifoliates 
```{r}

  df_MD <- df_trif %>%
  dplyr::select(-DAS50E)%>%
  mutate(Date50FirstTrif=SowingDate+DAS1T) %>%
  group_by(Plot, DAS1T) %>% 
  dplyr::filter(row_number() == 1)

head(df_MD)

df_MD <-df_MD %>%
 dplyr:: select (-Rate)

#save and store table 
#write.table(H, "DateToFirstTrifo.txt")

```

Read the soil Temp data with Max an Min records to calculate TT with the 8x3 hour formula 


```{r}

# Get dataframe 1 with the soil Temp in another folder 
setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\EmergenceIv2")
getwd()

#create file for the first trifo still using soil temperature (file SoilTempTXTM; if need air temperature use file RawweaIv in folder TrifoCounts Iv (5 Trifos) )
df_emerg6 <- read.table("SoilTempTXTM.txt",header=TRUE)
head(df_emerg6)

df_emerg6$Date <- dmy(df_emerg6$Date )

str(df_emerg6)
head(df_emerg6)

```

```{r}

#tell R to write in the folder first Trifo  

#write.txt(df_emerg6, "C:/GitHubRepos/SubCloverPhD/SubClover/FirstTrifo/SoilTempTXTM.txt")
write.table(df_emerg6, "C:/GitHubRepos/SubCloverPhD/SubClover/FirstTrifo/SoilTempTXTM.txt")

# then 




```




Add the accumulation of of thermal time use reference for all cultivars ! 


use consistent method as per re analysis tb=0, t opt range =15, 22, t max = 35 oC.



```{r}

#Create Thermal time cardinal temperatures see data above- TT = Topt - Tb ... 21.7 - (-1.85)= 23.55 ! 

#For cardinal temperatures 
#not in use here tt_card <- data.frame(temp=c(-1.9,15,22,35),TT=c(0,15,15,0))
#for Tb=0
tt_card <- data.frame(temp=c(0,15,22,35),TT=c(0,15,15,0))
tt_card %>%
  ggplot(aes(x=temp, y=TT))+
  #labs(x="Temperature (degree Celsius)")+
  #labs(y="Thermal Time (oCd)") +
  geom_line()+
  theme_bw()+
  ylim(0,20)+
xlab(expression(Temperature~"("^"o"*"C)"))+
  ylab(expression(Thermal~Time~Antas~"("^"o"*"Cd)"))+
  theme(axis.text.x=element_text(angle = 0, hjust = 0, size= 14))+
  theme(axis.title.x=element_text(angle = 0, hjust = 0.5, size= 14))+
  theme(axis.text.y=element_text(angle = 0, hjust = 0, size= 14))+
  theme(axis.title.y=element_text(size= 14))

```


 then make the loop
 
```{r}
#Create and test interpolation function - Test if works 

temp_ref  <- 25 # x-axes reference
temp <- tt_card$temp # x-axes
TT <- tt_card$TT # y-axes


int_func <- function(temp,TT,temp_ref){

  # if temp is too low or too high give extreme TT values
 if(temp_ref>temp[length(temp)]) {
   
   out <- TT[length(TT)] 
   # FIXME: approx functions seems to do that - need param
   
 } else if (temp_ref<temp[1]) {
   
   out <- TT[1]
   
 } else {
   
   # else interpolate
   
   tryCatch(  
   
  out <- approx(temp, TT, xout = temp_ref, 
         method="linear", 
         rule = 2)$y,
  error = function(e) 
  {
  out <- NA
  }
) 
  
} # end if check

  return(out)  
  
 }
  
int_func(temp,TT,temp_ref)


```
 

use the 3x8 hours formula to calculate daily TT 

```{r}

#create REF Jones CA, Ritchie JT, Kiniry JR, Godwin DC (1986) Subroutine structure.In 'CERES-Maize: A simulation model of maize growth anddevelopment'. (Eds CA Jones, JR Kiniry) pp. 49-194. (Texas A&M University Press: Texas, USA

TTfunc <- function(Tmin ,Tmax){         
 
  if (Tmin>Tmax){
   print("Tmin>Tmax")
   break
 } 
  
  TTav  <- 0
  TTsum <- 0
  
  for(n in 1:8){ 
    
    tn <- 0
    
    TT_tn <- 0
    
    tn <- ((0.931 +  0.114*n - 0.0703 * n^2 + 0.0053 * n^3) * (Tmax-Tmin)) + Tmin # 8-h temperature
    
    TT_tn <- int_func(tt_card$temp, tt_card$T,tn) 
    
    TTsum <- TT_tn + TTsum
    
   # print(TT_tn)
    
    }
 
   TTav <- TTsum/8
  
  return(TTav)
  }

TTfunc(10,24)


```

Then next step is 


- Compares with simple method to calculate Tt with base temperature of 0oC 
```{r, warning=FALSE}

df_We_tt <- df_emerg6 %>%
  #mutate(Date=dmy(Date)) %>%
  rowwise() %>%
  mutate(TTday=TTfunc(Tm,Tx), TTday_s=(Tm+Tx)/2, diff=(TTday_s-TTday))

summary(df_We_tt)

```



 
```{r}
#Check TT  two methods : TTday_s = simple method (Tmean - Tbzero)
df_We_tt  %>% 
ggplot(aes(x=TTday_s, y= TTday)) +
  geom_point(shape=21,alpha=0.2,colour='red')+
  geom_abline(intercept = 0, slope = 1)

```



 
 Then accumulate the TT 
 
 
```{r}

 df_FirstTrifo_TT <-df_We_tt%>%
    arrange(Date)%>%
    mutate(SumTTday = cumsum(TTday))%>%
  mutate(SUMTTday_s = cumsum(TTday_s)) %>%
  dplyr::select(-TempSoil10mm)

#head(df_soilTemp)

summary(df_FirstTrifo_TT)

#df_emerg6 <- NULL

```



figure out how to calculate tt at sow then tt at  date of 50% first trifo  make one minus another to find TT First Trifo 

```{r}
df1 <- merge(df_MD,df_FirstTrifo_TT, by.x="SowingDate",by.y="Date") %>%
  mutate(soilTempAcc_sow = SumTTday) %>%
  dplyr::select(-SumTTday)
 # filter(Cultivar=="Antas")

str(df1)
```



Need to force a date for First Trifo  (round) to calculate the average temperature during first trifo appearance 
Save dataframe as txt file 

```{r}

# force formats of dates
df1$Date <- round(df1$Date50FirstTrif,0)
df_FirstTrifo_TT$Date <- round(df_FirstTrifo_TT$Date,0)


  df_merge <- merge(df1, df_FirstTrifo_TT, by="Date") %>%
    mutate(soilTempAcc_em = SumTTday) %>%
  dplyr::select(-SumTTday) %>%
  mutate(SummTTFirstTrifo = (soilTempAcc_em-soilTempAcc_sow))  

summary(df_merge)

 
```


get the TT and SE for all cultivars. exclude S1 . Consider Tb=0 for all cultivars. 

```{r}
df_TT <- df_merge%>%
   filter(SowTreat!= "S1")

#then select sow treat and means and se

df_TTA <-df_TT %>%
  group_by(SowTreat, Cultivar) %>%
  dplyr::select( SummTTFirstTrifo)%>%
  summarise_all(funs(mean,sd))%>%
    mutate(SE=sd/1.73)
  
```


Then ANOVA 



## Create index to filter for loop ANOVA

```{r}
y <- c("S1","S2","S3","S4","S5","S6","S7","S8")

sdIndex <- unique(df_merge$SowTreat)

sdIndex[order(match(sdIndex,y))]

length(sdIndex)
sdIndex
```

# Then loop anova for each individual sowing date 

```{r}
for (i in 1:length(sdIndex)) {
  
 #print(i)
 print(paste0(sdIndex[i],"--------------------------------"))
  
  file.subset <- df_merge %>%
  filter(SowTreat==sdIndex[i])
  
 
head(file.subset, 50)
   summary(file.subset)
my.anova <- aov( SummTTFirstTrifo ~ Cultivar + Block, data = file.subset)
print(summary(my.anova))

if ((summary(my.anova)[[1]][["Pr(>F)"]][1])<0.05) {
  print("***Sig Cultivar***") 
  print((LSD.test(my.anova, c("Cultivar"), alpha= 0.05, p.adj="none"))) } 


if ((summary(my.anova)[[1]][["Pr(>F)"]][1])>0.05) {
  print("***Sig Cultivar***") 
  print((LSD.test(my.anova, c("Cultivar"), alpha= 0.05, p.adj="none"))) } 

# make mean separation if significant

(LSD.test(my.anova, c("Cultivar"), alpha= 0.05, p.adj="none"))
  
}

```
Now calculate DAS and TT from emergence to first trifoliate 
Then prepare dataframe 1 

```{r}

##Bring the table df_tt_50p_emerg

# Get dataframe 1 with the soil Temp in another folder 
setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\EmergenceIv2")
getwd()

#create file for the first trifo still using soil temperature (file SoilTempTXTM; if need air temperature use file RawweaIv in folder TrifoCounts Iv (5 Trifos) )
df_Emergence <- read.table("df_tt_50p_emerg.txt",header=TRUE)

df_Emergence<-df_Emergence%>%
  dplyr::select(Plot, DAS50E,TT50E)


str(df_Emergence)
head(df_Emergence)



```

Prepare dataframe 2 with DAS and TT of First trifo (df_merge). Need to select columns

```{r}
str(df_merge)


df_mergedSelect <- df_merge %>%
dplyr::select(SowTreat ,Cultivar,Plot,Block,DAS1T, Date50FirstTrif,SummTTFirstTrifo)


df_mergedFirstTrifo<-merge(df_mergedSelect,df_Emergence, by="Plot" )


```



Cool! Now subtract DASTrifo - DASE and TT50E - SummTTTrifo to get the values from emergence 



```{r}

str(df_mergedFirstTrifo)

df_mergedFirstTrifo <-df_mergedFirstTrifo%>%
mutate(DAS1TFromEmerg=DAS1T-DAS50E,
       TT1TFromEmerg=SummTTFirstTrifo-TT50E)

```

write a table with this dataset!!! 


```{r}

write.table(df_mergedFirstTrifo, "C:/GitHubRepos/SubCloverPhD/SubClover/FirstTrifo/DaysTT1TFromEmergence.txt")


```



End of this script . 
Move to script 7.  and make the stats for DAS and TT from emergence to first Trifoliate
